+++ 03_Template/Scene/EndNames/TextRendering.h
@@ -184,7 +184,7 @@ void initTextRendering()
 			texture,
 			ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			(GLuint)face->glyph->advance.x
 		};
 		Characters_TextRender.insert(std::pair<GLchar, Character>(c, character));
